PLC_Leas_Energy_JSON/requirements.txt
opcua
requests
python-dotenv
pytz
cryptography